Minor change to basicsubs.c.
Added millero to mytest/Makefile

Added millero, millero.csv, and millero.out to svn (not zero'd)

Updated revisions

git-svn-id: svn://136.177.114.72/svn_GW/phreeqc/trunk/src@3190 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ basicsubs.c
@@ -199,7 +199,7 @@ calc_dens (void)
        }
     }
 
-    z = abs(s_ptr->z);
+    z = abs((int) s_ptr->z);
     a = s_ptr->millero[0];
     b = s_ptr->millero[1];
     c = s_ptr->millero[2];
